Change metrics channel to use string
It seems that the metric channel was not always sending the complete datagram.
Changing this to a string seems to have resolved it but I end up having to
convert from bytes to string to pass it back to the main loop and from string to
bytes to hand it off to the socket. Not ideal but performance is still good.